+++ sfiaweb/sqlite-viewer-py/main.py
@@ -1,6 +1,0 @@
-def main():
-    print("Hello from sqlite-viewer-py!")
-
-
-if __name__ == "__main__":
-    main()